+++ lib/domain/usecases/teacher/get_recent_school_activity_usecase.dart
@@ -1,7 +1,6 @@
 import 'package:dartz/dartz.dart';
 
 import '../../../core/errors/failures.dart';
-import '../../entities/teacher.dart';
 import '../../repositories/teacher_repository.dart';
 import '../usecase.dart';